bugfix: explicitly return python floating point numbers from metrics functions
otherwise triggers some absurd problem involving numpy types

+++ gtorch/metrics/metrics.py
@@ -33,7 +33,7 @@ def next_token_metrics(model, val_loader, verbose=True):
                       index="t v_mag2 a_mag2 dv_mag2 cw j_mag2".split(),
                       name="Verbose MSE components"
                       ))
-  mse = np.mean((predicted - data)**2)
+  mse = float(np.mean((predicted - data)**2))
   return mse
 
 def binary_classifier_metrics(model, val_loader):
@@ -52,7 +52,7 @@ def binary_classifier_metrics(model, val_loader):
   # only needed for accuracy
   #predictions = np.argmax(logits, axis=1)
   targets = np.concatenate(targets)
-  return roc_auc_score(targets, logits[:, 1])
+  return float(roc_auc_score(targets, logits[:, 1]))
 
 def get_combined_roc(model, test_loader, combine_fn=None):
   logits = []
